Progression on UI. Moved "width" and "height" back to unsigned int.
git-svn-id: http://svn.ta3d.org/yuni/trunk@1239 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/gfx/surface/cairo.cpp
@@ -17,7 +17,7 @@ namespace Surface
 	void Cairo::release()
 	{}
 
-	void Cairo::resize(size_t, size_t)
+	void Cairo::resize(unsigned int, unsigned int)
 	{
 	}
